Move IO instances to `scalaz72` object and make them compile again
